Move Volt test dogfood fixtures under test
